refactor: package management, MatrixMemento, preps for Composite
