+++ jogo-oito/src/main/java/interfaces/Graph.java
@@ -4,6 +4,8 @@
  */
 package interfaces;
 
+import listeners.BoardUpdateListener;
+
 import java.util.List;
 
 /**
