+++ jogo-oito/src/main/java/listeners/BoardUpdateListener.java
@@ -1,4 +1,4 @@
-package interfaces;
+package listeners;
 
 public interface BoardUpdateListener {
