+++ jogo-oito/src/main/java/model/MatrixMemento.java
@@ -1,12 +1,12 @@
-package util;
+package model;
+
+import interfaces.Vertex;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import interfaces.Vertex;
-
-public record BoardMemento(List<Vertex> cells) {
-    public BoardMemento(List<Vertex> cells) {
+public record MatrixMemento(List<Vertex> cells) {
+    public MatrixMemento(List<Vertex> cells) {
         // Fazer cópias das células
         this.cells = new ArrayList<>();
         for (Vertex cell : cells) {
